refactor: segment-builder made modular
